Soft Swap & Splice
Added two new methods and adjusted code to support them. The first is a splice method that can be called on a linked list that returns a new linked list. The second is a "soft swap" that allows one to swap properties of two nodes. By swapping the data property (assuming this is the only uniquely identifying data of a node), one can effectively swap the two nodes' positions much more efficiently.